pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sugarai"
version = "3.4.5.dev2"
description = "🍰 Sugar - The autonomous layer for AI coding agents. Manages task queues, runs 24/7, ships working code."

readme = "README.md"
requires-python = ">=3.11"
license = {text = "AGPL-3.0-or-later"}
authors = [
    {name = "Steven Leggett", email = "contact@roboticforce.io"}
]
keywords = ["claude", "autonomous", "development", "ai", "automation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers", 
    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Quality Assurance"
]

dependencies = [
    "click>=8.1.0",
    "pyyaml>=6.0",
    "aiosqlite>=0.19.0",
    "sqlalchemy>=2.0.0",
    "python-dotenv>=1.0.0",
    "asyncio-throttle>=1.0.0",
    "setuptools>=45",
    "claude-agent-sdk>=0.1.0",
]

[project.optional-dependencies]
github = ["PyGithub>=1.59.0"]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "flake8-docstrings>=1.7.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "bandit>=1.7.5",
    "safety>=2.3.0",
    "pre-commit>=3.0.0",
    "types-PyYAML>=6.0.0",
    "types-requests>=2.28.0",
    "types-setuptools>=69.0.0"
]
test = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0"
]
mcp = [
    "mcp>=1.0.0",
    "starlette>=0.27.0",
    "uvicorn>=0.22.0",
]
memory = [
    "sentence-transformers>=2.2.0",
    "sqlite-vec>=0.1.0",
]
all = [
    "sugarai[github,mcp,memory]",
]

[project.scripts]
sugar = "sugar.main:cli"

[project.urls]
Homepage = "https://github.com/roboticforce/sugar"
Documentation = "https://sugar.roboticforce.io/"
Repository = "https://github.com/roboticforce/sugar"
"Bug Tracker" = "https://github.com/roboticforce/sugar/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["sugar*"]
exclude = ["tests*"]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
)
'''

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.bandit]
skips = ["B101", "B601"]

[tool.coverage.run]
source = ["sugar"]
omit = [
    "*/tests/*",
    "*/test_*",
    "setup.py",
    "venv/*",
    ".venv/*"
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod"
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = "-v --cov=sugar --cov-branch --cov-report=term-missing --cov-report=xml"
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow running tests"
]

# MCP Registry identification
# The mcpName must match the name in server.json for registry publishing
# See: https://github.com/modelcontextprotocol/registry
[tool.mcp]
mcpName = "io.github.cdnsteve/sugar"
